reorganize affine transform
