+++ dynamo/tools/pseudotime_velocity.py
@@ -208,3 +208,4 @@ def pseudotime_velocity(
         logger.info("set gamma to be 0 in .var. so that velocity_S = unspliced RNA.")
         logger.info_insert_adata("gamma", "var", indent_level=2)
         adata.var["gamma"] = 0
+        adata.var["gamma_b"] = 0
